bacon: name the package correctly
The variable CM_VERSION gets updated each build, but
the actual ro.cm.version in the build.prop doesn't on
incremental builds.

Name the 'bacon' package according to cm version
from the build.prop (if it exists), so that the actual
version and the name of the zip stay in sync.